split up ApplicaitonViews and NavBar into AdminViews/AuthorViews, AdminNavBar/AuthorNavBar
